+++ src/main/java/org/opentripplanner/routing/edgetype/LegSwitchingEdge.java
@@ -33,7 +33,7 @@ public class LegSwitchingEdge extends Edge {
 	}
 
 	@Override
-	public double getDistance() {
+	public double getDistanceMeters() {
 		return 0;
 	}