+++ src/main/java/org/opentripplanner/routing/edgetype/ParkAndRideEdge.java
@@ -74,7 +74,7 @@ public class ParkAndRideEdge extends Edge {
     }
 
     @Override
-    public double getDistance() {
+    public double getDistanceMeters() {
         return 0;
     }