+++ src/main/java/org/opentripplanner/routing/edgetype/StreetTransitLink.java
@@ -45,7 +45,7 @@ public class StreetTransitLink extends Edge {
         return null;
     }
 
-    public double getDistance() {
+    public double getDistanceMeters() {
         return 0;
     }